avoid too many open files for agent
Implementation remember the fd rlimit and stop accepting new connections when it is exceeded (with some grace). Accept is resumed when enough connections are closed.
This avoid situation when "ssh-agent starts eating 100% if it gets bombarded by connections, and runs out of file descriptors to use. ..."
It could be reproduced by example:
$ ulimit -n 12
$ for i in `seq 1 256`; do ssh localhost id & done
Note server has to be able to accept such number of concurrent connections.