pick "remove support for obsolete host/port syntax"
Host/port was added in 2001 as an alternative to host:port syntax for the benefit of IPv6 users.
These days there are establised standards for this like [::1]:22 and the slash syntax is easily mistaken for CIDR notation, which SSH now supports for some things.
Remove the slash notation from ListenAddress and PermitOpen.