correction for regression from "better diagnosics on alg list assembly errors"
Commit accidentally reused the server's algorithm lists in the client for KEX, ciphers and MACs.
The ciphers and MACs were identical between the client and server, but erroneously disable the diffie-hellman-group-exchange-sha1 KEX method for client.

This fixes the client code to use the correct lists.
As nobody complained, it does not restore diffie-hellman-group-exchange-sha1 KEX method for client.