pick "allow s390 specific ioctl for ecc hardware support"
Adding another s390 specific ioctl to be able to support ECC hardware acceleration to the sandbox seccomp filter rules.

Now the ibmca openssl engine provides elliptic curve cryptography support with the help of libica and CCA crypto cards. This is done via jet another ioctl call to the zcrypt device driver and so there is a need to enable this on the openssl sandbox.