adapt "add a sleep to allow forwards to come up" in multiplex test
Currently when the multiplex client requests a forward it returns once the request has been sent but not necessarily when the forward is up.
This causes intermittent test failures due to this race, so add some sleeps to mitigate this until we can fix it properly.

Remark: multiplex is excluded from regular regression tests