adapt additional "call comma-separated lists as such to clarify semantics" in client and daemon manual page
Options such as Ciphers take values that may be a list of ciphers; the complete list, not indiviual elements, may be prefixed with a minus or plus character to remove from or append to the default list, respectively.
Users might read the current text as if each elment took an optional prefix, so tweak the wording from "values" to "list" to prevent such ambiguity for all options supporting these semantics.