pick "re-add SA_RESTART to mysignal"
This makes mysignal implement reliable BSD semantics according to Stevens' APUE.
This was first attempted in 2001 but was reverted due to problems with HP-UX 10.20 and select() and possibly grantpt().
Modern systems should be fine with it, but if any current platforms have a problem with it now we can disable it just for those.