pick "fix shield/unshield for xmss keys"
Note xmss keys are not build by default.

In agent we need to delay the call to shield until we have received key specific options - when serializing xmss keys for shield we need to deal with all optional components (e.g. state might not be loaded).
In order to be able to figure out the number of signatures left on a shielded key, we need to transfer the number of signatures left from the private to the public key.