postpone build of certificate chain for agent keys
Usually agent keys are used with IdentityFile and when IdentityFile is loaded (it has private and public part) chain is build.

Unfortunately this does not work for PKCS#11 keys.
For such keys in case of RFC6187 algorithms if IdentityFile is set client send chain without intermediate certificates.

Instead to correct load of public identity, solution is for agent keys to postpone build of chain until chain is really needed.