prevent excessively long username going to PAM on Solaris
This is a mitigation for a buffer overflow in Solaris PAM username handling (CVE-2020-14871), and is only enabled for Sun-derived PAM implementations.
This is not a problem in secure shell daemon itself, it only prevents daemon from being used as a vector to attack Solaris PAM.
It does not prevent the bug in PAM from being exploited via some other PAM application.

(synchronisation)