in sftp command properly sort remote directory listing
Throws away timespeccmp - is result is boolean, i.e. not suitable for qsort(3).
Correct all comparison function to return qsort(3) expected values - an integer less than, equal to, or greater than zero if the first argument is considered to be respectively less than, equal to, or greater than the second.

Also use variables local to statement in comparison function.