do not log partial successes as failures
By design, 'partial' logins are successful logins, so initially with authenticated set to 1, for which another authentication is required.
As a result, authenticated is always reset to 0 when partial is set to 1.
However, even if authenticated is 0, those are not failed login attempts, similarly to attempts with authctxt->postponed set to 1.

(synchronisation)