add pselect compatible implementation
Remark:
This is basically the existing notify_pipe kludge from serverloop.c moved behind a pselect interface.
It works by installing a signal handler that writes to a pipe that the select is watching, then calls the original handler.

The select call in serverloop will become pselect soon, at which point the kludge will be removed from thereand will only exist in the own source.

(adaptation)