experimental handle interrupt on sftp "editline" related code
Remark: Require double Ctrl-C to cancel command line editing!

Gotal is user to use Ctrl-C to discard the currently edited command line and to get a fresh prompt.

For this particular case is critical to avoid use of ssl_signal() as it sets unconditionally SA_RESTART, but here need is signal to interrupt the read(2) in the el_gets(3) event loop.

(adaptation)