avoid race in pselect compatible implementation
On the second and subsequent calls to pselect the notify_pipe was not added to the select readset, opening up a race discovered on multiprocessor Solaris <=9 systems.
Also reinitialise notify_pipe if the pid changes. This will prevent a parent and child from using the same FD, although this is not an issue in the current structure it might be in future.

(adaptation)