switch hpdelim() to accept only ":" as delimiter
Historically, hpdelim() accepted ":" or "/" as a port delimiter between hosts (or addresses) and ports.
These days most of the uses for "/" are no longer accepted, so there are several places where it checks the delimiter to disallow it.
Make hpdelim() accept only ":" and use hpdelim2() in the other cases."

(adaptation)