return ERANGE from getcwd() compatible implementation if buffer size is 1
Quote "If getcwd() is supplied a buffer size of exactly 1 and a path of "/", it could result in a nul byte being written out of array bounds.  POSIX says it should return ERANGE if the path will not fit in the available buffer (with terminating nul). 1 byte cannot fit any possible path with its nul, so immediately return ERANGE in that case."

Remark: code does not use so small buffer, i.e. no impact.

(synchronisation)