rewrite "ensure that known hosts file ends with newline before to add a new entry"
Unfortunately adapted code relies on OpenBSD broken file stream management and does not work on OS that implement correctly POSIX requirement.

Quote from fopen specification: "When a file is opened with update mode ( '+' as the second or third character in the mode argument), both input and output may be performed on the associated stream. However, the application shall ensure that output is not directly followed by input without an intervening call to fflush() or to a file positioning function (fseek(), fsetpos(), or rewind()), and input is not directly followed by output without an intervening call to a file positioning function, unless the input operation encounters end-of-file."