release GSS OIDs only at end of authentication
Note: If the gssapi system reports multiple mechs it goes through the cycle init/cleanup for each mech. So the supported-mechs-list is on sort of a higher level than a single gssapi auth attempt. Hence, releasing the mechs-list itself should not be done in the gssapi_cleanup function but at the very end of authentication.

(adaptation)