refuse user or host names provided on the command line that contain most shell metacharacters
Some programs that invoke ssh client using untrusted data do not filter metacharacters in arguments they supply. This could create interactions with user-specified ProxyCommand and other directives that allow shell injection attacks to occur.

It's a mistake to invoke ssh client with arbitrary untrusted arguments, but getting this stuff right can be tricky, so this should prevent most obvious ways of creating risky situations. It however is not and cannot be perfect: commands has no practical way of interpreting what shell quoting rules are in use and how they interact with the user's specified ProxyCommand.

To allow configurations that use strange user or hostnames to continue to work, this strictness is applied only to names coming from the commandline. Names specified using User or Hostname directives in client configuration are not affected.

(synchronisation)